http.c: Honor Connection: close
Some HTTP clients will require the use of the "Connection: close" header
so that the server is forced to close the connection once the body is
sent. Otherwise, those HTTP clients will kept waiting for more data,
even if Content-Length is already sent by the server.

Field names other than "close" are ignored for backwards compatibility.